cleanup group entities
